refactor: new constants files were created and constants existing were moved
