+++ src/constants/icons.constants.ts
@@ -2,8 +2,8 @@ export { default as Add } from '../assets/Add.svg'
 export { default as AppGallery } from '../assets/app-galery.svg'
 export { default as AppStore } from '../assets/app-store.svg'
 export { default as ArchiveAdd } from '../assets/ArchiveAdd.svg'
-export { default as ArchiveTick } from '../assets/ArchiveTick.svg'
 export { default as ArchiveRemove } from '../assets/ArchiveRemove.svg'
+export { default as ArchiveTick } from '../assets/ArchiveTick.svg'
 export { default as ArrowDown2 } from '../assets/ArrowDown2.svg'
 export { default as ArrowLeft2 } from '../assets/ArrowLeft2.svg'
 export { default as ArrowRight2 } from '../assets/ArrowRight2.svg'
@@ -40,6 +40,7 @@ export { default as MailDark } from '../assets/Mail1.svg'
 export { default as MessageQuestion } from '../assets/MessageQuestion.svg'
 export { default as MinTrabajo } from '../assets/min-trabajo.svg'
 export { default as More } from '../assets/More1.svg'
+export { default as NoLogo } from '../assets/noLogo.svg'
 export { default as People } from '../assets/People.svg'
 export { default as ProfileAdd } from '../assets/ProfileAdd.svg'
 export { default as Refresh } from '../assets/Refresh.svg'
@@ -61,4 +62,3 @@ export { default as WhatsAppColor } from '../assets/Whatsapp1.svg'
 export { default as WhatsAppDark } from '../assets/Whatsapp.svg'
 export { default as X } from '../assets/X.svg'
 export { default as Youtube } from '../assets/Youtube.svg'
-export { default as NoLogo } from '../assets/noLogo.svg'
