--- src/constants/template/index.ts
@@ -1,2 +1,0 @@
-export * from './sideFilter.constants'
-export * from './sortBar.constants'
